'修改dev-runner'

+++ .electron-vite/dev-runner.ts
@@ -28,7 +28,7 @@ function logStats(proc: string, data: any) {
         data.toString({
             colors: true,
             chunks: false
-        }).split(/\r?\n/).forEach(line => {
+        }).split(/\r?\n/).forEach((line: string) => {
             log += '  ' + line + '\n'
         })
     } else {
@@ -87,9 +87,9 @@ function startMain(): Promise<void> {
         });
         MainWatcher.on('event', event => {
             if (event.code === 'END') {
-                if (electronProcess && electronProcess.kill) {
+                if (electronProcess) {
                     manualRestart = true
-                    process.kill(electronProcess.pid)
+                    if (electronProcess.pid) process.kill(electronProcess.pid)
                     electronProcess = null
                     startElectron()
 
@@ -116,18 +116,21 @@ function startElectron() {
     ]
 
     // detect yarn or npm and process commandline args accordingly
-    if (process.env.npm_execpath.endsWith('yarn.js')) {
+
+    console.log("process.env.npm_execpath", process.env.npm_execpath)
+
+    if (process.env.npm_execpath?.endsWith('yarn.js')) {
         args = args.concat(process.argv.slice(3))
-    } else if (process.env.npm_execpath.endsWith('npm-cli.js')) {
+    } else if (process.env.npm_execpath?.endsWith('npm-cli.js')) {
         args = args.concat(process.argv.slice(2))
     }
 
     electronProcess = spawn(electron as any, args)
 
-    electronProcess.stdout.on('data', (data: string) => {
+    electronProcess?.stdout?.on('data', (data: string) => {
         electronLog(removeJunk(data), 'blue')
     })
-    electronProcess.stderr.on('data', (data: string) => {
+    electronProcess?.stderr?.on('data', (data: string) => {
         electronLog(removeJunk(data), 'red')
     })
